+++ src/translations/en-GB/themes.php
@@ -1,6 +1,17 @@
 <?php 
 
 return [
-	'Default (No theme)' => 'Default (No theme)',
-	'Themes cache' => 'Themes cache'
+	'Themes rules' => 'Themes rules',
+	'Rules will be handled from top to bottom, the first match will define the theme and the other rules will not be looked at.' => 'Rules will be handled from top to bottom, the first match will define the theme and the other rules will not be looked at.',
+	'Default theme' => 'Default theme',
+	'Default theme when no rules match' => 'Default theme when no rules match',
+	'Themes cache' => 'Themes cache',
+	'Enabled' => 'Enabled',
+	'Path (or regex)' => 'Path (or regex)',
+	'Site' => 'Site',
+	'Any' => 'Any',
+	'Language' => 'Language',
+	'Theme' => 'Theme',
+	'No theme' => 'No theme',
+	'Enter path here' => 'Enter path here'
 ];