+++ src/translations/fr/themes.php
@@ -1,6 +1,17 @@
 <?php 
 
 return [
-	'Default (No theme)' => 'Défaut (Pas the thème)',
-	'Themes cache' => 'Cache des thèmes'
+	'Themes rules' => 'Règles de thèmes',
+	'Rules will be handled from top to bottom, the first match will define the theme and the other rules will not be looked at.' => 'Les règles sont traitées de haut en bas, la première règle valide definira le thème et les autres règles ne seront pas traitées.',
+	'Default theme' => 'Thème par défaut',
+	'Default theme when no rules match' => 'Thème par défaut quand aucune règle n\'est valide',
+	'Themes cache' => 'Cache des thèmes',
+	'Enabled' => 'Active',
+	'Path (or regex)' => 'Chemin (ou regex)',
+	'Site' => 'Site',
+	'Any' => 'Tous',
+	'Language' => 'Langage',
+	'Theme' => 'Thème',
+	'No theme' => 'Pas de thème',
+	'Enter path here' => 'Entrer le chemin ici'
 ];