+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # ryssbowh/craft-themes Changelog
 
+## 1.0.0 - 2021-01-31
+### Added
+- Rule system
+
+### Changed
+- Theme cache in vendor folder
+- Dependency to ryssbowh/craft-theme-installer-plugin
+- themes service renamed to registry
+
 ## 0.1.0 - 2021-01-27
 ### Added
 - Initial release